Move the logic to NetworkConnection file
