+++ src/libs/NetworkConnection.ts
@@ -1,4 +1,5 @@
 import NetInfo from '@react-native-community/netinfo';
+import {differenceInHours} from 'date-fns/differenceInHours';
 import isBoolean from 'lodash/isBoolean';
 import throttle from 'lodash/throttle';
 import Onyx from 'react-native-onyx';
@@ -6,6 +7,7 @@ import type {ValueOf} from 'type-fest';
 import CONFIG from '@src/CONFIG';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
+import type {ConnectionChanges} from '@src/types/onyx/Network';
 import * as NetworkActions from './actions/Network';
 import AppStateMonitor from './AppStateMonitor';
 import Log from './Log';
@@ -51,6 +53,7 @@ const triggerReconnectionCallbacks = throttle(
  * then all of the reconnection callbacks are triggered
  */
 function setOfflineStatus(isCurrentlyOffline: boolean, reason = ''): void {
+    trackConnectionChanges();
     NetworkActions.setIsOffline(isCurrentlyOffline, reason);
 
     // When reconnecting, ie, going from offline to online, all the reconnection callbacks
@@ -64,12 +67,25 @@ function setOfflineStatus(isCurrentlyOffline: boolean, reason = ''): void {
 
 // Update the offline status in response to changes in shouldForceOffline
 let shouldForceOffline = false;
+let isPoorConnectionSimulated: boolean | undefined;
+let connectionChanges: ConnectionChanges | undefined;
 Onyx.connect({
     key: ONYXKEYS.NETWORK,
     callback: (network) => {
         if (!network) {
             return;
         }
+
+        // Starts random network status change when shouldSimulatePoorConnection is turned into true
+        // or after app restart if shouldSimulatePoorConnection is true already
+        if (!isPoorConnectionSimulated && !!network.shouldSimulatePoorConnection) {
+            clearTimeout(network.poorConnectionTimeoutID);
+            setRandomNetworkStatus(true);
+        }
+
+        isPoorConnectionSimulated = !!network.shouldSimulatePoorConnection;
+        connectionChanges = network.connectionChanges;
+
         const currentShouldForceOffline = !!network.shouldForceOffline;
         if (currentShouldForceOffline === shouldForceOffline) {
             return;
@@ -103,6 +119,59 @@ Onyx.connect({
         accountID = session.accountID;
     },
 });
+
+/** Sets online/offline connection randomly every 2-5 seconds */
+function setRandomNetworkStatus(initialCall = false) {
+    // The check to ensure no new timeouts are scheduled after poor connection simulation is stopped
+    if (!isPoorConnectionSimulated && !initialCall) {
+        setOfflineStatus(false);
+        return;
+    }
+
+    const statuses = [CONST.NETWORK.NETWORK_STATUS.OFFLINE, CONST.NETWORK.NETWORK_STATUS.ONLINE];
+    const randomStatus = statuses[Math.floor(Math.random() * statuses.length)];
+    const randomInterval = Math.random() * (5000 - 2000) + 2000; // random interval between 2-5 seconds
+    Log.info(`[NetworkConnection] Set connection status "${randomStatus}" for ${randomInterval} sec`);
+
+    setOfflineStatus(randomStatus === CONST.NETWORK.NETWORK_STATUS.OFFLINE);
+
+    const timeoutID = setTimeout(setRandomNetworkStatus, randomInterval);
+    NetworkActions.setPoorConnectionTimeoutID(timeoutID);
+}
+
+function simulatePoorConnection(shouldSimulatePoorConnection: boolean, poorConnectionTimeoutID: NodeJS.Timeout | undefined) {
+    if (!shouldSimulatePoorConnection) {
+        clearTimeout(poorConnectionTimeoutID);
+        NetworkActions.setPoorConnectionTimeoutID(undefined);
+        setOfflineStatus(false);
+    }
+
+    NetworkActions.setShouldSimulatePoorConnection(shouldSimulatePoorConnection);
+}
+
+/** Tracks how many times the connection has changed within the time period */
+function trackConnectionChanges() {
+    if (!connectionChanges?.startTime) {
+        NetworkActions.setConnectionChanges({startTime: new Date().getTime(), amount: 1});
+        return;
+    }
+
+    const diffInHours = differenceInHours(new Date(), connectionChanges.startTime);
+    const newAmount = (connectionChanges.amount ?? 0) + 1;
+
+    if (diffInHours < 1) {
+        NetworkActions.setConnectionChanges({amount: newAmount});
+        return;
+    }
+
+    Log.info(
+        `[NetworkConnection] Connection has changed ${newAmount} time(s) for the last ${diffInHours} hour(s). Poor connection simulation is turned ${
+            isPoorConnectionSimulated ? 'on' : 'off'
+        }`,
+    );
+
+    NetworkActions.setConnectionChanges({startTime: new Date().getTime(), amount: 0});
+}
 
 /**
  * Set up the event listener for NetInfo to tell whether the user has
@@ -227,5 +296,6 @@ export default {
     triggerReconnectionCallbacks,
     recheckNetworkConnection,
     subscribeToNetInfo,
+    simulatePoorConnection,
 };
 export type {NetworkStatus};
